experiments/ai-research-collection/requirements.txt
# Core ML/AI Dependencies
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.0.0

# Progress and CLI
tqdm>=4.65.0
rich>=13.0.0

# Configuration
pyyaml>=6.0

# Web frameworks (for optional GUI)
flask>=2.3.0
gradio>=3.40.0

# Image processing
Pillow>=10.0.0

# Environment management
python-dotenv>=1.0.0

# OpenAI integration
openai>=1.0.0

# TinyTroupe (if available via pip, otherwise needs manual install)
# tinytroupe

# Markdown processing
markdown>=3.4.0